Improved: the package structure to move the api directory to src(#2h1fjk7)
