+++ index.ts
@@ -1,7 +1,7 @@
 import { getOrderDetails, updateOrderStatus } from './src/order'
 import { events } from './src/types'
 import { fetchProducts } from './src/product'
-import { init, updateToken, updateInstanceUrl } from './api'
+import { init, updateToken, updateInstanceUrl } from './src/api'
 
 export {
   getOrderDetails,
